pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sai", email = "ysaijagadeeshuh@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "azure-communication-email>=1.1.0",
    "azure-core>=1.36.0",
    "azure-data-tables>=12.7.0",
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.27.1",
    "deltalake>=1.2.1",
    "dotenv>=0.9.9",
    "duckdb>=1.4.2",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "playwright>=1.55.0",
    "polars>=1.34.0",
    "pyarrow>=21.0.0",
    "pydantic-settings>=2.13.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
    "xlsxwriter>=3.2.9",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=7.2.0",
]
